requirements.txt
Django==4.2.23
djangorestframework==3.16.1
uvicorn==0.35.0
channels==4.3.1
channels_redis==4.3.0
whitenoise==6.9.0
django-redis==6.0.0
psycopg2==2.9.10
django-environ==0.12.0
pillow==11.3.0
django-ckeditor-5==0.2.18
celery==5.5.3
django-celery-results==2.6.0